"Uncommented test cases and added easy level tag"
Uncommented the calls to the longestCommonPrefix function with different input arrays in the Longest_Common_Prefix_14.java file to ensure these test cases are part of the testing process. Also, added a level tag to the Length_of_Last_Word_58.java file to specify the difficulty level of the problem.