Add Palindrome Number solution
A new solution file for the problem "Palindrome Number" has been added. This problem was categorized as "Easy" and the included solution had a runtime of 9 ms and a memory use of 42.60 mb. The solution includes two methods: isPalindrome() and isPalindrome2(), both of which take an integer as input and return a boolean indicating whether the input is a palindrome. The time complexity of both methods is O(n).